feat(backend)(security)(workflow)(zap): add secret-gated scan fixture endpoint for zap af state setup
