+++ apps/backend/src/main/resources/application.yml
@@ -4,3 +4,8 @@ spring:
 
 server:
   port: 8080
+
+app:
+  scan-fixture:
+    enabled: ${APP_SCAN_FIXTURE_ENABLED:false}
+    secret: ${APP_SCAN_FIXTURE_SECRET:}
