+++ apps/backend/src/test/resources/application.yml
@@ -1,3 +1,3 @@
 app:
-  scan-fixtures:
+  scan-fixture:
     enabled: false
